Implement temperature unit converter.

+++ app/src/main/java/me/claudiuconstantinbogdan/weatherapp/util/temperature/TemperatureConverter.java
@@ -1,6 +1,6 @@
 package me.claudiuconstantinbogdan.weatherapp.util.temperature;
 
 public interface TemperatureConverter {
-    double convertTemperature(double fahrenheitTemperature);
+    int convertTemperature(double fahrenheitTemperature);
     String getTemperatureUnits();
 }
